EventPipeline/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure-storage-blob
azure-ai-documentintelligence
azure-cosmos
azure-ai-contentunderstanding
azure-identity
redis
python-dotenv
